solution-es-grafana/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
autopep8==2.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
elastic-transport==8.4.0
elasticsearch==8.5.3
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
numpy==1.23.5
pandas==1.5.2
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.13
yarl==1.8.2
